HR Chatbot/backend/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.3.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asn1crypto==1.5.1
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
atpublic==5.0
attrs==24.2.0
azure-common==1.1.28
azure-core==1.31.0
azure-data-tables==12.5.0
azure-functions==1.21.3
azure-functions-durable==1.2.9
azure-monitor-opentelemetry-exporter==1.0.0b25
azure-search-documents==11.5.1
azure-storage-blob==12.23.1
Babel==2.15.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
botocore==1.35.38
cachetools==5.3.3
captcha==0.6.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==2.2.1
comm==0.2.2
cryptography==43.0.3
dataclasses-json==0.6.4
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
distro==1.9.0
exceptiongroup==1.2.0
executing==2.0.1
extra-streamlit-components==0.1.71
faiss-cpu==1.8.0
fastjsonschema==2.20.0
filelock==3.13.3
fixedint==0.1.6
Flask==2.3.3
Flask-Cors==4.0.1
Flask-Injector==0.15.0
fqdn==1.5.1
frozenlist==1.4.1
furl==2.1.3
gitdb==4.0.11
GitPython==3.1.42
grpcio==1.53.0
grpcio-tools==1.53.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.10
importlib-metadata==7.0.0
injector==0.21.0
ipykernel==6.29.5
ipython==8.26.0
isodate==0.7.2
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
jq==1.7.0
json5==0.9.25
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
langchain==0.1.12
langchain-community==0.0.28
langchain-core==0.1.31
langchain-experimental==0.0.54
langchain-openai==0.0.8
langchain-text-splitters==0.0.1
langsmith==0.1.24
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mmh3==4.1.0
msrest==0.7.1
multidict==6.1.0
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.2.2
openai==1.13.3
opentelemetry-api==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-flask==0.45b0
opentelemetry-instrumentation-requests==0.45b0
opentelemetry-instrumentation-wsgi==0.45b0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
orderedmultidict==1.0.1
orjson==3.9.15
overrides==7.7.0
packaging==23.2
pandas==2.2.1
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==10.2.0
platformdirs==3.11.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
propcache==0.2.0
protobuf==4.25.3
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==15.0.1
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
pydeck==0.8.1b0
Pygments==2.17.2
PyJWT==2.8.0
pymilvus==2.2.5
pymssql==2.3.1
pyOpenSSL==20.0.1
pypdf==5.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.33.0
regex==2023.12.25
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.0
s3transfer==0.10.1
Send2Trash==1.8.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
snowflake==0.11.0
snowflake-connector-python==3.12.2
snowflake-snowpark-python==1.20.0
snowflake-sqlalchemy==1.5.1
snowflake._legacy==0.11.0
snowflake.core==0.11.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==1.4.52
stack-data==0.6.3
streamlit==1.32.1
streamlit-authenticator==0.3.3
tenacity==8.2.3
terminado==0.18.1
tiktoken==0.6.0
tinycss2==1.3.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.4
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
types-python-dateutil==2.9.0.20240316
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.9.0
uri-template==1.3.0
urllib3==2.2.3
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.4
wrapt==1.16.0
yarl==1.15.5
zipp==3.20.2
gunicorn==23.0.0
pytesseract==0.3.13
azure-storage-blob==12.23.1
flasgger==0.9.7.1
